+++ crates/nostr/src/event/tag/mod.rs
@@ -18,12 +18,14 @@ use serde::de::Error as DeserializerError;
 use serde::ser::SerializeSeq;
 use serde::{Deserialize, Deserializer, Serialize, Serializer};
 
+pub mod cow;
 mod error;
 pub mod kind;
 pub mod list;
 pub mod standard;
 pub(super) mod weak;
 
+pub use self::cow::CowTag;
 pub use self::error::Error;
 pub use self::kind::TagKind;
 pub use self::list::Tags;